// src/generated/api/models/PagedResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 모든 API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0.4
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
/**
 *
 * @export
 * @interface PagedResponse
 */
export interface PagedResponse {
  /**
   *
   * @type {Array<object>}
   * @memberof PagedResponse
   */
  content?: Array<object>;
  /**
   *
   * @type {number}
   * @memberof PagedResponse
   */
  page?: number;
  /**
   *
   * @type {number}
   * @memberof PagedResponse
   */
  size?: number;
  /**
   *
   * @type {number}
   * @memberof PagedResponse
   */
  totalElements?: number;
  /**
   *
   * @type {number}
   * @memberof PagedResponse
   */
  totalPages?: number;
}

/**
 * Check if a given object implements the PagedResponse interface.
 */
export function instanceOfPagedResponse(value: object): value is PagedResponse {
  return true;
}

export function PagedResponseFromJSON(json: any): PagedResponse {
  return PagedResponseFromJSONTyped(json, false);
}

export function PagedResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): PagedResponse {
  if (json == null) {
    return json;
  }
  return {
    content: json["content"] == null ? undefined : json["content"],
    page: json["page"] == null ? undefined : json["page"],
    size: json["size"] == null ? undefined : json["size"],
    totalElements: json["totalElements"] == null ? undefined : json["totalElements"],
    totalPages: json["totalPages"] == null ? undefined : json["totalPages"],
  };
}

export function PagedResponseToJSON(json: any): PagedResponse {
  return PagedResponseToJSONTyped(json, false);
}

export function PagedResponseToJSONTyped(
  value?: PagedResponse | null,
  ignoreDiscriminator: boolean = false
): any {
  if (value == null) {
    return value;
  }

  return {
    content: value["content"],
    page: value["page"],
    size: value["size"],
    totalElements: value["totalElements"],
    totalPages: value["totalPages"],
  };
}